Rework Thread Work Allocation Scheme
Threads now take multiples of boids based on 64-byte cache lines. This follows the data-oriented practice. The boid group size is set to 8. 64 * 8 is a multiple of all the 2D 16-byte layout, the 3D 24-byte layout, and the quaternion 32-byte layout.

Special care is taken to evenly distribute work at any boid count.